requirements.txt
fastmcp>=0.1.0
pyyaml>=6.0
mistune>=3.0.0